requirements.txt
tweepy>=3.7.0
Pillow>=6.0.0
requests>=2.21.0